Split sub-API classes into separate modules (#28)
 * Keeps the API compatibility
